+++ Biblioteca.Aplicacion/Docente/CasosDeUsoDocente/ModificarDocenteUseCase.cs
@@ -1,0 +1,15 @@
+namespace Biblioteca.Aplicacion;  
+
+public class ModificarDocenteUseCase   
+{
+    private readonly IRepositorioDocente _repo;
+    public ModificarDocenteUseCase(IRepositorioDocente repo)
+    {
+        this._repo = repo;
+    }
+    public void Ejecutar(Docente docente)
+    {
+        _repo.modificarDocente(docente);
+    }
+    
+}
