+++ Biblioteca.Aplicacion/Estudiante/CasosDeUsoEstudiante/BajaEstudiante.cs
@@ -1,0 +1,17 @@
+namespace Biblioteca.Aplicacion;
+
+
+
+public class BajaEstudianteUseCase
+{
+    private readonly IRepositorioEstudiante _repo;
+    public BajaEstudianteUseCase(IRepositorioEstudiante repo)
+    {
+        this._repo = repo;
+    }
+    public void Ejecutar(int idEstudiante)
+    {
+        _repo.bajaEstudiante(idEstudiante);
+    }
+    
+}
