+++ Biblioteca.Aplicacion/Libro/CasosDeUsoLibro/ListarLibrosUseCase.cs
@@ -1,0 +1,17 @@
+namespace Biblioteca.Aplicacion;
+
+
+public class ListarLibrosUseCase
+{
+    private readonly IRepositorioLibro _repo;
+    public ListarLibrosUseCase(IRepositorioLibro repo)
+    {
+        this._repo = repo;
+    }
+
+    public List<Libro> Ejecutar()
+    {
+        return _repo.listarLibros();
+    }
+
+}
